requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.90.0
pytest-mock>=3.12.0
